Create DelayedIO module
