+++ hs-bindgen/app/HsBindgen/Cli/Info/BuiltinMacros.hs
@@ -64,7 +64,7 @@ parseOpts = Opts <$> parseClangArgsConfig
 
 exec :: GlobalOpts -> Opts -> IO ()
 exec GlobalOpts{..} Opts{..} =
-    void . withTracer tracerConfig $ \tracer _ -> do
+    void . withTracer tracerConfig $ \tracer -> do
       let tracerBoot  = contramap TraceBoot tracer
           tracerClang = contramap (TraceFrontend . FrontendClang) tracer
       (clangArgs, _target) <- getClangArgsAndTarget tracerBoot clangArgsConfig
