+++ hs-bindgen/app/HsBindgen/Cli/Preprocess.hs
@@ -17,9 +17,9 @@ import Options.Applicative hiding (info)
 
 import HsBindgen
 import HsBindgen.App
-import HsBindgen.Artefact
 import HsBindgen.Config
 import HsBindgen.Config.Internal
+import HsBindgen.DelayedIO
 import HsBindgen.Frontend.RootHeader
 import HsBindgen.Imports
